+++ internal/cli/dialog/proposal_breadcrumb.go
@@ -35,8 +35,12 @@ func ProposalBreadcrumb(args Args[forgedomain.ProposalBreadcrumb]) (Option[forge
 			Text: "no breadcrumb in proposals, or use the Git Town GitHub Action",
 		},
 		{
-			Data: Some(forgedomain.ProposalBreadcrumbCLI),
-			Text: "Git Town CLI embeds the breadcrumb into proposals",
+			Data: Some(forgedomain.ProposalBreadcrumbStacks),
+			Text: "Git Town CLI embeds the breadcrumbs for stacks containing more than 2 branches into proposals",
+		},
+		{
+			Data: Some(forgedomain.ProposalBreadcrumbBranches),
+			Text: "Git Town CLI embeds the breadcrumbs into all proposals",
 		},
 	}...)
 	defaultPos := entries.IndexOf(args.Local)