+++ internal/forge/forgedomain/proposal_breadcrumb_test.go
@@ -26,8 +26,13 @@ func TestParseProposalBreadcrumb(t *testing.T) {
 			err:  false,
 		},
 		{
-			give: "cli",
-			want: Some(forgedomain.ProposalBreadcrumbCLI),
+			give: "branches",
+			want: Some(forgedomain.ProposalBreadcrumbBranches),
+			err:  false,
+		},
+		{
+			give: "stacks",
+			want: Some(forgedomain.ProposalBreadcrumbStacks),
 			err:  false,
 		},
 		{
@@ -37,7 +42,7 @@ func TestParseProposalBreadcrumb(t *testing.T) {
 		},
 		{
 			give: "true",
-			want: Some(forgedomain.ProposalBreadcrumbCLI),
+			want: Some(forgedomain.ProposalBreadcrumbBranches),
 			err:  false,
 		},
 		{
@@ -47,7 +52,7 @@ func TestParseProposalBreadcrumb(t *testing.T) {
 		},
 		{
 			give: "yes",
-			want: Some(forgedomain.ProposalBreadcrumbCLI),
+			want: Some(forgedomain.ProposalBreadcrumbBranches),
 			err:  false,
 		},
 		{
@@ -57,7 +62,7 @@ func TestParseProposalBreadcrumb(t *testing.T) {
 		},
 		{
 			give: "1",
-			want: Some(forgedomain.ProposalBreadcrumbCLI),
+			want: Some(forgedomain.ProposalBreadcrumbBranches),
 			err:  false,
 		},
 		{